@@ -1,28 +1,18 @@
 source 'http://rubygems.org'
 
-gem 'rails', '3.2.12'
-# Bundle edge Rails instead:
-# gem 'rails', :git => 'git://github.com/rails/rails.git'
+gem 'rails', '4.0.0'
 
-gem 'sqlite3-ruby', :require => 'sqlite3'
-
+gem 'sqlite3-ruby'
 gem 'devise'
-group :assets do
-  gem 'sass-rails',   '~> 3.2.3'
-  gem 'coffee-rails', '~> 3.2.1'
-  gem 'uglifier',     '>= 1.0.3'
-end
 
-# Bundle gems for the local environment. Make sure to
-# put test-only gems in this group so their generators
-# and rake tasks are available in development mode:
-group :development do
-  gem 'annotate-models'
 
-end
+gem 'sass-rails', '~> 4.0.0'  
+gem 'coffee-rails', '~> 4.0.0'
+gem 'uglifier', '>= 1.3.0'
+gem 'jquery-rails'
 
-group :test do
 
-end
+
+
 
 
