@@ -1,4 +1,5 @@
 class AuthorsController < ApplicationController
+  before_action :set_author, only: [:show, :update, :destroy]
   # GET /authors
   # GET /authors.xml
   def index
@@ -9,7 +10,6 @@ class AuthorsController < ApplicationController
   # GET /authors/1
   # GET /authors/1.xml
   def show
-    @author = Author.find(params[:id])
     respond_with(@author)
   end
 
@@ -28,7 +28,7 @@ class AuthorsController < ApplicationController
   # POST /authors
   # POST /authors.xml
   def create
-    @author = Author.new(params[:author])
+    @author = Author.new(author_params)
     @author.save
     respond_with(@author)
   end
@@ -36,16 +36,24 @@ class AuthorsController < ApplicationController
   # PUT /authors/1
   # PUT /authors/1.xml
   def update
-    @author = Author.find(params[:id])
-    @author.update_attributes(params[:author])
+    @author.update(author_params)
     respond_with(@author)
   end
 
   # DELETE /authors/1
   # DELETE /authors/1.xml
   def destroy
-    @author = Author.find(params[:id])
     @author.destroy
-    respond_with(@author)
+    redirect_to authors_path
+  end
+  
+  private
+  
+  def author_params
+    params.require(:author).permit(:name, :age)
+  end
+  
+  def set_author
+    @author = Author.find(params[:id])
   end
 end
