@@ -1,8 +1,3 @@
 # Be sure to restart your server when you modify this file.
 
-Bloga::Application.config.session_store :cookie_store, :key => '_bloga_session'
-
-# Use the database for sessions instead of the cookie-based default,
-# which shouldn't be used to store highly confidential information
-# (create the session table with "rake db:sessions:create")
-# Bloga::Application.config.session_store :active_record_store
+Bloga::Application.config.session_store :cookie_store, key: '_bloga_session'
