@@ -1,39 +1,26 @@
-# This file is auto-generated from the current state of the database. Instead 
+# encoding: UTF-8
+# This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
 #
-# Note that this schema.rb definition is the authoritative source for your 
+# Note that this schema.rb definition is the authoritative source for your
 # database schema. If you need to create the application database on another
 # system, you should be using db:schema:load, not running all the migrations
 # from scratch. The latter is a flawed and unsustainable approach (the more migrations
 # you'll amass, the slower it'll run and the greater likelihood for issues).
 #
-# It's strongly recommended to check this file into your version control system.
+# It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(:version => 20100914160927) do
+ActiveRecord::Schema.define(version: 20100914151847) do
 
-  create_table "authors", :force => true do |t|
+  create_table "authors", force: true do |t|
     t.string   "name"
     t.integer  "age"
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.string   "email",                               :default => "", :null => false
-    t.string   "encrypted_password",   :limit => 128, :default => "", :null => false
-    t.string   "password_salt",                       :default => "", :null => false
-    t.string   "reset_password_token"
-    t.string   "remember_token"
-    t.datetime "remember_created_at"
-    t.integer  "sign_in_count",                       :default => 0
-    t.datetime "current_sign_in_at"
-    t.datetime "last_sign_in_at"
-    t.string   "current_sign_in_ip"
-    t.string   "last_sign_in_ip"
   end
 
-  add_index "authors", ["email"], :name => "index_authors_on_email", :unique => true
-  add_index "authors", ["reset_password_token"], :name => "index_authors_on_reset_password_token", :unique => true
-
-  create_table "comments", :force => true do |t|
+  create_table "comments", force: true do |t|
     t.integer  "post_id"
     t.string   "author"
     t.text     "body"
@@ -41,7 +28,7 @@ ActiveRecord::Schema.define(:version => 20100914160927) do
     t.datetime "updated_at"
   end
 
-  create_table "posts", :force => true do |t|
+  create_table "posts", force: true do |t|
     t.string   "title"
     t.text     "body"
     t.datetime "created_at"
