@@ -1,7 +1,5 @@
-require "application_responder"
-
 class ApplicationController < ActionController::Base
-  self.responder = ApplicationResponder
+ 
   respond_to :html
 
   protect_from_forgery
