@@ -1,7 +1,12 @@
 # Be sure to restart your server when you modify this file.
 
-# Your secret key for verifying the integrity of signed cookies.
+# Your secret key is used for verifying the integrity of signed cookies.
 # If you change this key, all old signed cookies will become invalid!
+
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-Bloga::Application.config.secret_token = '8a4a5821f90fea4d3d1d22dadd4ada4a961fa0858efb527f3d50b19d35dc45b3860074d340922d2a8bd2dda322b690afb6eca579deafdc3eebbeb700d008a7a7'
+# You can use `rake secret` to generate a secure secret key.
+
+# Make sure your secret_key_base is kept private
+# if you're sharing your code publicly.
+Bloga::Application.config.secret_key_base = 'fbe4a70646eaafac64af4fba4e8c19bfa9012561686e533b7f93f2e7df33821e586c7fb0fc3b53e3ddf9eafcfb78cea4bc3bd83cec8992693246bab4b75252ed'
